# docs/control-plane/openapi.yaml
openapi: 3.0.3
info:
  title: NOFX Control Plane API
  description: |
    The NOFX Control Plane orchestration API for managing execution runs, steps, and artifacts.

    ## Overview
    The control plane turns human or system requests into durable, auditable execution runs.
    It normalizes incoming plans, persists them, schedules work on the queue, tracks lineage,
    and streams updates to any surface (UI, CLI, Slack bots, downstream services).

    ## Authentication
    Most endpoints require admin authentication via cookie. Use `/login` endpoint to authenticate.

    ## Base URL
    Default: `http://localhost:3000`
  version: 1.0.0
  contact:
    name: NOFX Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.nofx.io
    description: Production server (placeholder)
...
paths:
...
  /runs:
    post:
      tags:
        - Runs
      summary: Create a new run
      description: |
        Create a new execution run from a plan or standard prompt.

        Supports two modes:
        1. **Plan mode**: Provide a structured plan with steps
        2. **Standard mode**: Provide a natural language prompt that will be converted to a plan
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required:
                    - plan
                  properties:
                    plan:
                      $ref: '#/components/schemas/Plan'
                    projectId:
                      type: string
                      default: default
                      description: Project ID to associate with the run
                - type: object
                  required:
                    - standard
                  properties:
                    standard:
                      type: object
                      required:
                        - prompt
                      properties:
                        prompt:
                          type: string
                          description: Natural language prompt describing the task
                        quality:
                          type: boolean
                          default: true
                          description: Enable quality gates
                        openPr:
                          type: boolean
                          default: false
                          description: Open a pull request after completion
                        filePath:
                          type: string
                          description: Target file path for operations
                        summarizeQuery:
                          type: string
                          description: Query for summarization
                        summarizeTarget:
                          type: string
                          description: Target for summarization
            examples:
              planMode:
                summary: Create run with structured plan
                value:
                  plan:
                    goal: "Write documentation"
                    steps:
                      - name: "write_readme"
                        tool: "codegen"
                        inputs:
                          topic: "API Documentation"
                          content: "Write comprehensive API docs"
              standardMode:
                summary: Create run from prompt
                value:
                  standard:
                    prompt: "Write unit tests for the auth module"
                    quality: true
                    openPr: false
      responses:
        '201':
          description: Run created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - status
                properties:
                  id:
                    type: string
                    format: uuid
                    description: Unique run ID
                  status:
                    type: string
                    enum: [queued]
                  projectId:
                    type: string
                    description: Associated project ID
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Runs
      summary: List runs
      description: Get a list of recent runs with optional filtering
      operationId: listRuns
      parameters:
        - name: limit
          in: query
          description: Maximum number of runs to return
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: projectId
          in: query
          description: Filter by project ID
          schema:
            type: string
      responses:
        '200':
          description: List of runs
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Run'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nofx_admin
  schemas:
    Plan:
      type: object
      required:
        - goal
        - steps
      properties:
        goal:
          type: string
          description: High-level goal of the execution plan
        steps:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Step'
    Step:
      type: object
      required:
        - name
        - tool
      properties:
        name:
          type: string
          description: Step name/identifier
        tool:
          type: string
          description: Tool to execute (e.g., codegen, gate:typecheck, manual:approve)
        inputs:
          type: object
          additionalProperties: true
          description: Tool-specific input parameters
        tools_allowed:
          type: array
          items:
            type: string
          description: Allowed tools for security policy
        env_allowed:
          type: array
          items:
            type: string
          description: Allowed environment variables
        secrets_scope:
          type: string
          description: Secrets access scope
    Run:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
        project_id:
          type: string
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
        goal:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
...